+++ src/rotater.c
@@ -389,7 +389,7 @@ static int zlog_rotater_parse_archive_path(zlog_rotater_t * a_rotater)
 
 		nwrite = snprintf(a_rotater->glob_path + len, sizeof(a_rotater->glob_path) - len,
 				"*%s", p + nread + 1);
-		if (nwrite < 0 || nwrite > sizeof(a_rotater->glob_path) - len) {
+		if (nwrite < 0 || nwrite > (int)(sizeof(a_rotater->glob_path) - len)) {
 			zc_error("nwrite[%d], overflow or errno[%d]", nwrite, errno);
 			return -1;
 		}
